introduce the assign operation, operate and clear function
BUG SPOTTED: The render function renders on screen the last_number variable, which after passing through the Operate function it doesn't actually sets itself to the value but to the second_number, since its original functionality was to perform as a memory holder rather than the result holder.

POSSIBLE FIX: Introduce a result variable.